`btap_datapoint.rb`, `btap_data.rb`, `btap_analysis.rb`: Renaming these and moving them to `btap/qaqc/`
